Add files via upload
This is a complete working database registration application built in Java using the postgresQL database. It uses the Java SpringBoot framework to create the login information and authenticates the user via custom token. It will in the future include Facebook and Google+ login options using OKTA SSO, and will be deployed into Amazon Cognito server for remote access and storage.